refactor: use double for price, create Product List screen

+++ lib/models/product.dart
@@ -16,7 +16,7 @@ class Product {
       id: product['id'],
       title: product['title'],
       image: product['thumbnail'],
-      price: product['price'],
+      price: product['price'].toDouble(),
     );
   }
 }
